Some changes for error handling

+++ files/user.js
@@ -20,6 +20,9 @@ async function get(id)
     let result  = await database.getUser('id', id);
     let user    = [];
 
+    if (result.length === 0)
+        throw Error("This user doesn't exist");
+
     user.push({id:result[0].id, name:result[0].name});
 
     return user;
@@ -29,27 +32,38 @@ async function add(requestBody)
 {
     let name    = requestBody.name;
     let exist   = await checkName(name);
+
     if (!exist)
     {
         await database.addUser(name);
-        return true;
+        return "User "+name+" added to the base!";
     }
+    throw Error("This user already exist");
 }
 
 async function update(id, requestBody)
 {
     let name    = requestBody.name;
     let exist   = await checkName(name);
-    if (!exist)
+    let checkId = await database.getUser('id', id);
+
+    if (!exist && checkId.length > 0)
     {
         await database.updateUser({name:name, id:id});
-        return true;
+        return "User "+checkId[0].name+" has been updated to "+name;
     }
+    throw Error ("Either the name is already in use or that user doesn't exists");
 }
 
 async function deleteUser(id)
 {
-    return await database.deleteUser(id);
+    let checkId = await database.getUser('id', id);
+    if (checkId.length > 0)
+    {
+        await database.deleteUser(id);
+        return "User "+checkId[0].name+" has been deleted";
+    }
+    throw Error ("That user doesn't exists");
 }
 
 module.exports = { list, get, add, update, deleteUser };
